+++ src/core/lang/verilog/primaryunit.rs
@@ -16,6 +16,7 @@
 //
 
 use super::{symbols::VerilogSymbol, token::identifier::Identifier};
+use crate::core::lang;
 use crate::core::lang::vhdl::primaryunit::HdlNamingError;
 use crate::{core::lang::verilog::symbols::VerilogParser, util::anyerror::CodeFault};
 use std::collections::HashMap;
@@ -131,10 +132,7 @@ pub fn collect_units(files: &Vec<String>) -> Result<HashMap<Identifier, PrimaryU
         if crate::core::fileset::is_verilog(&source_file) == true {
             // println!("parse verilog: {:?}", source_file);
             // parse text into Verilog symbols
-            let contents = match std::fs::read_to_string(&source_file) {
-                Ok(dump) => dump,
-                Err(e) => return Err(CodeFault(Some(source_file.clone()), Box::new(e))),
-            };
+            let contents = lang::read_to_string(&source_file)?;
             let symbols = match VerilogParser::read(&contents) {
                 Ok(s) => s.into_symbols(),
                 Err(e) => Err(CodeFault(Some(source_file.clone()), Box::new(e)))?,