+++ src/core/lang/vhdl/primaryunit.rs
@@ -18,6 +18,7 @@
 use super::super::lexer::Position;
 use super::subunit::SubUnit;
 use super::symbols::VhdlSymbol;
+use crate::core::lang;
 use crate::core::lang::reference::RefSet;
 use crate::core::lang::vhdl::symbols::VHDLParser;
 use crate::core::lang::vhdl::token::identifier::Identifier;
@@ -161,10 +162,7 @@ pub fn collect_units(files: &Vec<String>) -> Result<HashMap<Identifier, PrimaryU
         if crate::core::fileset::is_vhdl(&source_file) == true {
             // parse text into VHDL symbols
             // println!("parsing vhdl: {}", &source_file);
-            let contents = match std::fs::read_to_string(&source_file) {
-                Ok(dump) => dump,
-                Err(e) => return Err(CodeFault(Some(source_file.clone()), Box::new(e))),
-            };
+            let contents = lang::read_to_string(&source_file)?;
             let symbols = match VHDLParser::read(&contents) {
                 Ok(s) => s.into_symbols(),
                 Err(e) => Err(CodeFault(Some(source_file.clone()), Box::new(e)))?,